requirements.txt
async-timeout==3.0.1
blinker==1.4
brotlipy==0.7.0
certifi==2020.12.5
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
dataclasses==0.7
future==0.18.2
gym==0.17.3
halo==0.0.31
inflection==0.5.1
jsonpatch==1.28
jsonpointer==2.0
kiwisolver==1.3.1
log-symbols==0.0.14
matplotlib==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.6.0
numpy
oauthlib==3.1.0
olefile==0.46
pandas==1.1.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
PyJWT==1.7.1
pyparsing==2.4.7
PySide2==5.15.2
python-chess==0.31.4
python-dateutil==2.8.1
pytz==2020.4
pyzmq==20.0.0
qt-thread-updater==1.1.4
QtPy==1.9.0
Quandl==3.5.3
scipy==1.5.4
shiboken2==5.15.2
spinners==0.0.24
stockfish==1.0.4
tensorboard-plugin-wit==1.6.0
termcolor==1.1.0
torch==1.7.1
torchaudio==0.7.2
torchfile==0.1.0
torchvision==0.8.2
tornado==6.1
urllib3==1.26.2
visdom==0.1.8.9
websocket-client==0.57.0
Werkzeug==1.0.1
wincertstore==0.2

